// emp-mgm/src/app/model/roles.model.ts
export class Roles {
    _id?: string;
    name: string;
    permission: Rolespermission;
    editable: boolean;
}